media: v4l2: Add NV15 pixel format
Add the NV15 pixel format used by the Rockchip Video Decoder for 10-bit buffers.

NV15 is a packed 10-bit 4:2:0 Y/CbCr format similar to P010 and P210 but has no
padding between components. Instead, luminance and chrominance samples are grouped
into 4s so that each group is packed into an integer number of bytes:

YYYY = UVUV = 4 * 10 bits = 40 bits = 5 bytes

The '15' suffix refers to the optimum effective bits per pixel
which is achieved when the total number of luminance samples is a multiple
of 8 for NV15.

Signed-off-by: Jonas Karlman <jonas@kwiboo.se>
Signed-off-by: Sebastian Fricke <sebastian.fricke@collabora.com>
[cherry-pick and add a dummy M variant]
Signed-off-by: Hsia-Jun(Randy) Li <randy.li@synaptics.com>

Change-Id: Ia83bca5221099a4e9264454738f3bde2e80e4463
Reviewed-on: https://sc-debu-git.synaptics.com/gerrit/195035
Reviewed-by: Randy Li <Randy.Li@synaptics.com>
Reviewed-by: Yanyan Peng <Yanyan.Peng@synaptics.com>
Reviewed-by: Kaicheng Xie <Kaicheng.Xie@synaptics.com>
Reviewed-on: https://gerrit-sha.synaptics.com/gerrit/202749